Replace file and sed for filename pattern match
Previously the script relied on using the file command and its
platform specific database of magic numbers to find .pem files.
This worked on Debian and presumably others, but broke on the
default file command/database on OS X.

We replace with a simple file suffix pattern match. All AXA pem
files are installed with the .pem extension so this should work on
all platforms.