fix many typos and some punctuation and capitalization
This fixes varous mispellings.

It adds some closing sentence periods (to be consistent with
neighboring formatting).

It capitalizes SSH to be consistent with other uses in same doc
when referencing the protocol (not the tool).

Fixes mistake where suggests "make" is in current directory.
Removed "$ " from a prompt to be consistent with command formatting
in neighboring sentences.

This commit should no have any run-time code changes.