fix openssl thread locks
when the apikey transport was added, we were sharing a mutex array amongst both transports which effectively divided in half the number of available locks. this patch fixes that and fixes a memory leak that was a second order effect of twice allocating the same lock array using the same pointer.