add libedit wchar (Unicode) support
libedit as of 3.11 decided to force Unicode support by changing the callback signature of the `getcfn` (`el_rfunc_t`) callback from a `char *` to a `wchar_t *`. We handle that by reading a wchar and converting to a char. Note this patch does not port AXA to support Unicode, only to correctly handle editline/readline library variants vis-a-vis Unicode/`getcfn`.